Add db-testkit-doobie module. Rewrite old test in terms of PureharmRowTest <3
